+++ R/QIM_diabetes.R
@@ -507,7 +507,8 @@ list_qim_diabetes_appointments <- function(dMeasureQIM_obj,
 #'  if not supplied, reads $qim_ignoreOld
 #' @param lazy recalculate the diabetes contact list?
 #'
-#' @return dataframe of Patient (name), demographics, measure (done or not), InternalID, Count, proportion
+#' @return dataframe of Patient (name), demographics, measure (done or not), InternalID, Count, Proportion,
+#'  Proportion_demographic
 #' @export
 report_qim_diabetes <- function(dMeasureQIM_obj,
                                 contact = NA,
@@ -628,7 +629,10 @@ report_qim_diabetes <- function(dMeasureQIM_obj,
         dplyr::select(., intersect(names(.), c(report_groups, "n")))
       } %>>%
       # if no rows, then grouping will not remove unnecessary columns
-      dplyr::mutate(Proportion = prop.table(n))
+      dplyr::mutate(Proportion = prop.table(n)) %>>%
+      dplyr::group_by_at(demographic) %>>%
+      dplyr::mutate(Proportion_Demographic = prop.table(n)) %>>%
+      dplyr::ungroup()
     # proportion (an alternative would be proportion = n / sum(n))
 
     if (self$dM$Log) {